+++ src/main/java/blusunrize/immersiveengineering/api/utils/shapes/CachedVoxelShapes.java
@@ -1,12 +1,10 @@
-package blusunrize.immersiveengineering.common.util.shapes;
+package blusunrize.immersiveengineering.api.utils.shapes;
 
-import net.minecraft.util.Direction;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.shapes.IBooleanFunction;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.util.math.shapes.VoxelShapes;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;