+++ changelog.md
@@ -1,6 +1,12 @@
 ##### Version 1.15.2-3.2.1-pre
 - Fixed broken rendering of components on the Automated Workbench (BluSunrize)
 
+##### Version 1.14.4-3.2.1-pre
+- Changed the pump to be configured for input on the bottom by default (Malte)
+- Changed nugget to ingot/ingot to block recipes to only trigger when an IE ingot is used in the middle slot (Malte)
+  - If more mods adopt this convention it will mean that crafting a storage block from a specific mod will be possible
+- Fixed continued depletion of already depleted mineral veins (BluSunrize)
+
 ##### Version 1.15.2-3.2.0-114
 - Fixed mirrored machines rendering incorrectly (Malte)
 - Fixed bucket wheels in mirrored excavators rendering incorrectly (Malte)